+++ util/tracing/src/lib.rs
@@ -24,6 +24,7 @@ mod config;
 mod error;
 
 pub use self::config::Config;
+pub use self::error::fail_span;
 pub use self::error::Error;
 pub use self::error::ErrorKind;
 pub use self::error::Result;